added -s option to keep the generated source

+++ process.sh
@@ -1,16 +1,25 @@
 #!/bin/bash
 
+OPTSTRING=":s" #option to keep the generated source
+
 if [ "$#" -ne 1 ];then
 	echo "usage: process <script>"
 	
 	
 	
 else
+
+    getopts ${OPTSTRING} opt
+
+
+
     ./wrapper $1 temp.cpp
     g++ -o temp temp.cpp
     ./temp
 
-    rm temp.cpp
     rm temp
 
+    if [${opt} -ne "s"]; then
+        rm temp.cpp
+    fi
 fi
